+++ client/cache/error_code.go
@@ -128,8 +128,18 @@ Invalid load function for read-through pattern decorator.
 You should pass a valid(non-nil) load function when initiate the decorator instance.
 `)
 
-var LoadFuncFailed = berror.DefineCode(4002024, moduleName, "InvalidLoadFunc", `
+var LoadFuncFailed = berror.DefineCode(4002024, moduleName, "LoadFuncFailed", `
 Failed to load data, please check whether the loadfunc is correct
+`)
+
+var InvalidInitParameters = berror.DefineCode(4002025, moduleName, "InvalidInitParameters", `
+Invalid init cache parameters.
+You can check the related function to confirm that if you pass correct parameters or configure to initiate a Cache instance.
+`)
+
+var PersistCacheFailed = berror.DefineCode(4002026, moduleName, "PersistCacheFailed", `
+Failed to execute the StoreFunc.
+Please check the log to make sure the StoreFunc works for the specific key and value.
 `)
 
 var DeleteFileCacheItemFailed = berror.DefineCode(5002001, moduleName, "DeleteFileCacheItemFailed", `